Fix typo

+++ src/model/frame.rs
@@ -3,15 +3,22 @@ use super::data::Data;
 #[derive(Default, Clone, Debug)]
 pub struct Frame {
     pub auto: bool,
-    pub shift: f64,
-    pub min_x: f64,
+    pub shift: usize,
+    pub min_x: usize,
     pub min_y: f64,
-    pub max_x: f64,
+    pub max_x: usize,
     pub max_y: f64,
 }
 
 impl Frame {
-    pub fn new(auto: bool, shift: f64, min_x: f64, min_y: f64, max_x: f64, max_y: f64) -> Frame {
+    pub fn new(
+        auto: bool,
+        shift: usize,
+        min_x: usize,
+        min_y: f64,
+        max_x: usize,
+        max_y: f64,
+    ) -> Frame {
         Frame {
             auto,
             shift,
@@ -26,11 +33,11 @@ impl Frame {
         self.auto = auto;
     }
 
-    pub fn set_shift(&mut self, shift: f64) {
+    pub fn set_shift(&mut self, shift: usize) {
         self.shift = shift;
     }
 
-    pub fn set_min_x(&mut self, min_x: f64) {
+    pub fn set_min_x(&mut self, min_x: usize) {
         self.min_x = min_x;
     }
 
@@ -38,7 +45,7 @@ impl Frame {
         self.min_y = min_y;
     }
 
-    pub fn set_max_x(&mut self, max_x: f64) {
+    pub fn set_max_x(&mut self, max_x: usize) {
         self.max_x = max_x;
     }
 
@@ -76,8 +83,8 @@ impl Frame {
 
     fn auto_frame_x(&mut self, data: &[Data]) {
         let range = self.max_x - self.min_x;
-        self.min_x = data.len() as f64 - range + self.shift;
-        self.max_x = data.len() as f64 + self.shift;
+        self.min_x = data.len() - range + self.shift;
+        self.max_x = data.len() + self.shift;
     }
 }
 
@@ -87,24 +94,24 @@ mod frame_tests {
 
     #[test]
     fn test_frame() {
-        let mut frame = Frame::new(false, 0.0, 0.0, 0.0, 0.0, 0.0);
+        let mut frame = Frame::new(false, 0, 0, 0.0, 0, 0.0);
         frame.set_auto(true);
-        frame.set_shift(5.0);
-        frame.set_min_x(1.0);
+        frame.set_shift(5);
+        frame.set_min_x(1);
         frame.set_min_y(2.0);
-        frame.set_max_x(3.0);
+        frame.set_max_x(3);
         frame.set_max_y(4.0);
         assert_eq!(frame.auto, true);
-        assert_eq!(frame.shift, 5.0);
-        assert_eq!(frame.min_x, 1.0);
+        assert_eq!(frame.shift, 5);
+        assert_eq!(frame.min_x, 1);
         assert_eq!(frame.min_y, 2.0);
-        assert_eq!(frame.max_x, 3.0);
+        assert_eq!(frame.max_x, 3);
         assert_eq!(frame.max_y, 4.0);
     }
 
     #[test]
     fn test_auto_frame_x() {
-        let mut frame = Frame::new(true, 2.0, 0.0, 0.0, 5.0, 0.0);
+        let mut frame = Frame::new(true, 2, 0, 0.0, 5, 0.0);
         let mut data_list: Vec<Data> = Vec::new();
         data_list.push(Data::new(1, 0.0, 1.0, 1.0, 0.0, 0.0));
         data_list.push(Data::new(2, 0.0, 2.0, 2.0, 0.0, 0.0));
@@ -112,13 +119,13 @@ mod frame_tests {
         data_list.push(Data::new(4, 0.0, 4.0, 4.0, 0.0, 0.0));
         data_list.push(Data::new(5, 0.0, 5.0, 5.0, 0.0, 0.0));
         frame.auto_frame_x(&data_list);
-        assert_eq!(frame.min_x, 2.0);
-        assert_eq!(frame.max_x, 7.0);
+        assert_eq!(frame.min_x, 2);
+        assert_eq!(frame.max_x, 7);
     }
 
     #[test]
     fn test_auto_frame_y() {
-        let mut frame = Frame::new(true, 0.0, 0.0, 0.0, 5.0, 0.0);
+        let mut frame = Frame::new(true, 0, 0, 0.0, 5, 0.0);
         let mut data_list: Vec<Data> = Vec::new();
         data_list.push(Data::new(1, 0.0, 1.0, 1.0, 0.0, 0.0));
         data_list.push(Data::new(2, 0.0, 2.0, 2.0, 0.0, 0.0));
@@ -132,7 +139,7 @@ mod frame_tests {
 
     #[test]
     fn test_auto_frame() {
-        let mut frame = Frame::new(true, 2.0, 0.0, 0.0, 5.0, 0.0);
+        let mut frame = Frame::new(true, 2, 0, 0.0, 5, 0.0);
         let mut data_list: Vec<Data> = Vec::new();
         data_list.push(Data::new(1, 0.0, 1.0, 1.0, 0.0, 0.0));
         data_list.push(Data::new(2, 0.0, 2.0, 2.0, 0.0, 0.0));
@@ -140,8 +147,8 @@ mod frame_tests {
         data_list.push(Data::new(4, 0.0, 4.0, 4.0, 0.0, 0.0));
         data_list.push(Data::new(5, 0.0, 5.0, 5.0, 0.0, 0.0));
         frame.auto_frame(&data_list);
-        assert_eq!(frame.min_x, 2.0);
-        assert_eq!(frame.max_x, 7.0);
+        assert_eq!(frame.min_x, 2);
+        assert_eq!(frame.max_x, 7);
         assert_eq!(frame.min_y, 3.0);
         assert_eq!(frame.max_y, 5.0);
     }
